add ring builder for ring snapshot

+++ src/main/java/io/github/NK8916/HashFunction.java
@@ -3,5 +3,5 @@ package io.github.NK8916;
 import java.math.BigInteger;
 
 public interface HashFunction {
-    BigInteger hash(String key);
+    long hash(String key);
 }
